Fix __init__.py for filesystem_model and opendap_model (+tests)
